+++ src/main/java/joshie/harvest/plugins/crafttweaker/CraftTweaker.java
@@ -1,6 +1,8 @@
 package joshie.harvest.plugins.crafttweaker;
 
+import joshie.harvest.core.commands.CommandManager;
 import joshie.harvest.core.util.HFLoader;
+import joshie.harvest.shops.command.HFCommandShops;
 import minetweaker.MineTweakerAPI;
 
 @HFLoader(mods = "MineTweaker3")
@@ -8,5 +10,6 @@ public class CraftTweaker {
     public static void init() {
         MineTweakerAPI.registerClass(Shipping.class);
         MineTweakerAPI.registerClass(Shops.class);
+        CommandManager.INSTANCE.registerCommand(new HFCommandShops());
     }
 }