+++ src/main/resources/changelog.md
@@ -2,6 +2,7 @@
 * Add config option for giving the cheat buildings instead of blueprints
 * Add a hardy turnip crop, that takes three days to grow and can grow in any season (Only available in the tutorial)
 * Add tooltip to animal spawners
+* Added command to dump list of shops for CraftTweaker purposes (/hf shops)
 * Change the colour of leaves/grass for all seasons
 * Reverse the logic for crops being shippable. This may make existing crops unsellable.
 * Jade will trade shears for sickles, along with the buckets/hoes up until you get the blacksmith